+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+# sdaccel-builder v0.17.1
+
+## Features
+
+## Bugfixes
+
+# sdaccel-builder v0.17.0
+
+## Features
+
+## Bugfixes
+
+# sdaccel-builder v0.16.4
+
+## Features
+
+## Bugfixes
+
+# sdaccel-builder v0.16.3
+
+## Features
+
+## Bugfixes
+
+# sdaccel-builder v0.16.2
+
+## Features
+
+## Bugfixes
+
+# sdaccel-builder v0.16.1
+
+## Features
+
+## Bugfixes
+
+# sdaccel-builder v0.16.0
+
+## Features
+
+## Bugfixes
+
 # sdaccel-builder v0.15.0
 
 ## Features